requirements.txt
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-flask==1.3.0
setuptools==70.1.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
virtualenv==20.26.3
Werkzeug==3.0.3
